perf(model): use input linear rather than ff
the reasoning here is quite simple: we already have quite a few linear layers inside of the model. especially for video-only models it doesn't make a difference whether the linear layers are inside or outside of the model.
however, for video + token the thing is a bit different. in this case I believe that the memory freed from removing these operations can be better utilized by adding more layers to the reversible blocks.